Move selected filters summary to global builders
